% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateElectricityDemand.R
\name{calcElectricityDemand}
\alias{calcElectricityDemand}
\title{Calculate electricity demand}
\usage{
calcElectricityDemand()
}
\value{
A \code{\link[magclass:magclass-package]{magpie}} object.
}
\description{
Calculate electricity demand
}
\author{
Falk Benke
}
