% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertPLEXOS.R
\name{convertPLEXOS}
\alias{convertPLEXOS}
\title{Convert PLEXOS data}
\usage{
convertPLEXOS(x)
}
\arguments{
\item{x}{A \code{\link[magclass:magclass-package]{magpie}} object returned from
\code{\link[=readPLEXOS]{readPLEXOS()}}.}
}
\value{
A \code{\link[magclass:magclass-package]{magpie}} object.
}
\description{
Convert PLEXOS data
}
\author{
Falk Benke
}
