% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrdieter-package.R
\docType{package}
\name{mrdieter-package}
\alias{mrdieter}
\alias{mrdieter-package}
\title{mrdieter: MadRat DIETER Input Data Package}
\description{
The mrdieter package contains data preprocessing for the DIETER model.
}
\author{
\strong{Maintainer}: Falk Benke \email{benke@pik-potsdam.de}

}
