% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPLEXOS.R
\name{readPLEXOS}
\alias{readPLEXOS}
\title{Read PLEXOS World 2015 Electricity Demand}
\usage{
readPLEXOS()
}
\value{
A \code{\link[magclass:magclass-package]{magpie}} object.
}
\description{
Read PLEXOS World 2015 Electricity Demand
}
\author{
Falk Benke
}
